DERBY-5036: fix impossible comparison due to sign extension
Patch contributed by Dave Brosius <dbrosius@apache.org>.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1070676 13f79535-47bb-0310-9956-ffa450edef68

+++ java/drda/org/apache/derby/impl/drda/DRDAConnThread.java
@@ -3428,7 +3428,7 @@ class DRDAConnThread extends Thread {
 					break;
 				case CodePoint.TRGDFTRT:
 					byte b = reader.readByte();
-					if (b == 0xF1)
+					if (b == (byte)0xF1)
 						database.sendTRGDFTRT = true;
 					break;
 				//optional - not used in JCC so skip for now
